Add file statuses lambda
This lambda will take the input from the backend checks and the redacted
files lambda and return a list of file statuses.

We aren't returning any consignment stautuses at the moment because the
only ones we're currently setting are set outside the backend checks so
don't make a lot of sense here.

This connects to the database to get the allowed and disallowed PUIDs.
I'll remove the status check file statuses from the API and deploy that
when this is deployed.