This lambda is now running in a step function.
There is a size limit of 256Kb for messages passed between steps and for large consignments, the message is more than this.

To solve this, each step reads its input from S3 and once it has carried out its step, writes the new output to S3 again.

I've updated this lambda to use the backend-check-utils classes and
updated the test to work with calls to S3.