Fix status issues with multiple PUIDs
There are a few different fixes here.

The first is that if a file has more than one PUID, we were generating a
file status row for each PUID instead of one row for both. If one PUID
was allowed and one wasn't, the consignment status was showing as
Completed because one of the file statuses was a success. This now
checks all PUIDs and if any of them fail, the status is set to the
failure reason.

The second thing is that the list of successful PUIDs for the
ClientChecks status was reporting a success if *any* of the client
checks succeeded. This has been changed so it only reports success if
there are no failures at all.

The last is a change to the tests. We weren't checking what would happen
if ZeroByteFile was active in the DisallowedPuids table. It isn't in
prod at the moment but it could be switched on so I've added in another
test.