Add Failed status for missing data.
The backend checks have been changed so if there is an error with the
lambda, the processing continues but the file check result for that file
is missing.

I've updated the statuses to show Failed if the relevant file check is
missing and added some tests which I think cover the cases.